requirements.txt
requests==2.11.1
PyYAML==3.12